+++ src/main/java/ai/mcpdirect/gateway/dao/mapper/aitool/MCPToolMapper.java
@@ -3,7 +3,7 @@ package ai.mcpdirect.gateway.dao.mapper.aitool;
 import org.apache.ibatis.annotations.*;
 
 @Mapper
-public interface AIToolMapper extends ToolAgentMapper,ToolPermissionMapper,ToolMakerMapper,TeamToolMakerMapper {
+public interface MCPToolMapper extends MCPToolAgentMapper, MCPToolPermissionMapper, MCPToolMakerMapper, MCPTeamToolMakerMapper {
     String TABLE_NAME = "aitool.tool";
     String TABLE_NAME_V = "aitool.virtual_tool";
 }